This commit refs #541, update algorithm and filter add compare by word algorithm

+++ libs/modules/analysis/diff/diff.h
@@ -1,15 +1,7 @@
-/*--------------------------------------------------------------------*/
-/*		     Epitech EIP 2017 groupe Copypeste		      */
-/*								      */
-/*			    Algo diff				      */
-/*								      */
-/* @by :	Guillaume Krier					      */
-/* @created :	02/02/2015					      */
-/* @update :	25/02/2015					      */
-/*--------------------------------------------------------------------*/
+
 #pragma once
 
-/*\* PROTOTYPES *\*/
+/* PROTOTYPES */
 
 /*
 ** This function allows to get the difference value between two files 
@@ -18,4 +10,4 @@
 ** @param: char *str_file2 - string of lines file two
 ** @return: Integer - return comparaison value
 */
-float diff(char *str_file1, char *str_file2);
+double diff(char *str_file1, char *str_file2);
